refactor: move all enums in a single file
