+++ src/event_sourcing/dto/events/user/user_created.py
@@ -2,8 +2,7 @@ from pydantic import Field
 
 from event_sourcing.dto.base import ModelConfigBaseModel
 from event_sourcing.dto.events.base import EventDTO
-from event_sourcing.enums import EventType, Role
-from event_sourcing.infrastructure.enums import HashingMethod
+from event_sourcing.enums import EventType, HashingMethod, Role
 
 
 class UserCreatedDataV1(ModelConfigBaseModel):
