--- src/event_sourcing/infrastructure/enums.py
@@ -1,9 +1,0 @@
-"""Infrastructure-specific enums."""
-
-from enum import Enum
-
-
-class HashingMethod(str, Enum):
-    """Available hashing methods for passwords."""
-
-    BCRYPT = "bcrypt"
