+++ tests/integration/infrastructure/test_event_store_read_model_integration.py
@@ -16,8 +16,7 @@ from event_sourcing.dto.events.base import EventDTO
 from event_sourcing.dto.events.user.user_created import UserCreatedDataV1
 from event_sourcing.dto.events.user.user_updated import UserUpdatedDataV1
 from event_sourcing.dto.user import UserReadModelData
-from event_sourcing.enums import AggregateTypeEnum, EventType
-from event_sourcing.infrastructure.enums import HashingMethod
+from event_sourcing.enums import AggregateTypeEnum, EventType, HashingMethod
 
 if TYPE_CHECKING:
     from event_sourcing.infrastructure.event_store.psql import (
