+++ tests/unit/application/queries/handlers/user/test_get_user_history.py
@@ -12,8 +12,7 @@ from event_sourcing.application.queries.user.get_user_history import (
 )
 from event_sourcing.dto.events.factory import EventFactory
 from event_sourcing.dto.user import UserDTO
-from event_sourcing.enums import AggregateTypeEnum
-from event_sourcing.infrastructure.enums import HashingMethod
+from event_sourcing.enums import AggregateTypeEnum, HashingMethod
 
 
 @pytest.fixture
